[BACKLOG-18970] - Parquet Output - Add new options - Date/Time Stamp & File extension - fix option tab controls
